Renamed ProductVariationsMapper to ProductVariationsBulkUpdateMapper and moved it to seperate file
